Added pages: adding a precinct, floor.

+++ ticket/views.py
@@ -1,50 +1,99 @@
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.mixins import LoginRequiredMixin
+from django.db.models import Count
 from django.http import JsonResponse, request, HttpResponse
 from django.shortcuts import render, get_object_or_404
 from django.urls import reverse_lazy
 from django.views.generic import CreateView, ListView, UpdateView, DeleteView
 import openpyxl
 from ticket.filters import TicketFilter
-from ticket.forms import CreateTicketForm
+from ticket.forms import CreateTicketForm, CreatePrecinctForm, CreateExamForm, FloorCreateForm
 from ticket.models import Ticket, Floor, Room, ExamType, Precinct, Exam
 import logging
 
 logger = logging.getLogger(__name__)
 
 
-class TicketListView(ListView):
+class TicketListView(LoginRequiredMixin, ListView):
     model = Ticket
     template_name = 'ticket/index.html'
     context_object_name = 'tickets'
 
     def get_queryset(self):
-        queryset = super().get_queryset()
+        queryset = Ticket.objects.filter(user=self.request.user)  # Фильтруем билеты по пользователю
         self.filter = TicketFilter(self.request.GET, queryset=queryset)
         return self.filter.qs
 
     def get_context_data(self, *, object_list=None, **kwargs):
         context = super().get_context_data(**kwargs)
         context['filter'] = self.filter
+        context['ticket_counter'] = Ticket.objects.filter(user=self.request.user).count()  # Считаем только свои билеты
         return context
 
 
+class CreateTicketView(LoginRequiredMixin, CreateView):
+    model = Ticket
+    form_class = CreateTicketForm
+    template_name = 'ticket/add-ticket.html'
+    success_url = reverse_lazy('ticket:index')
+
+    def get_form_kwargs(self):
+        """Передаем текущего пользователя в форму"""
+        kwargs = super().get_form_kwargs()
+        kwargs['user'] = self.request.user  # Передаем user
+        return kwargs
+
+    def form_valid(self, form):
+        """Привязываем билет к пользователю"""
+        form.instance.user = self.request.user
+        return super().form_valid(form)
+
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data()
+        return context
+
+
+class TicketUpdate(LoginRequiredMixin, UpdateView):
+    model = Ticket
+    form_class = CreateTicketForm
+    template_name = 'ticket/update.html'
+    success_url = reverse_lazy('ticket:update')
+
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+        ticket = self.object
+
+        context['form'].fields['exam_type'].initial = ticket.exam_type
+        context['form'].fields['precinct'].initial = ticket.precinct
+        context['form'].fields['floor'].initial = ticket.floor
+        context['form'].fields['room'].initial = ticket.room
+        context['form'].fields['seat_number'].initial = ticket.seat_number
+
+        return context
+
+
+@login_required
 def load_exam_type(request):
     exam = request.GET.get('exam')
     exam_typs = ExamType.objects.filter(exam=exam).values('id', 'name')
     return JsonResponse(list(exam_typs), safe=False)
 
 
+@login_required
 def load_precinct(request):
     exam = request.GET.get('exam')
     precincts = Precinct.objects.filter(exam=exam).values('id', 'name')
     return JsonResponse(list(precincts), safe=False)
 
 
+@login_required
 def load_floor(request):
     precinct_id = request.GET.get('precinct_id')
     floors = Floor.objects.filter(precinct_id=precinct_id).values('id', 'name')
     return JsonResponse(list(floors), safe=False)
 
 
+@login_required
 def load_room(request):
     floor_id = request.GET.get('floor_id')
     exam_type_id = request.GET.get('exam_type')
@@ -62,6 +111,7 @@ def load_room(request):
     return JsonResponse(list(rooms), safe=False)
 
 
+@login_required
 def load_seats(request):
     room_id = request.GET.get('room_id')
 
@@ -82,43 +132,7 @@ def load_seats(request):
         return JsonResponse({'error': 'Internal Server Error'}, status=500)
 
 
-class CreateTicketView(CreateView):
-    model = Ticket
-    form_class = CreateTicketForm
-    template_name = 'ticket/add-ticket.html'
-    success_url = reverse_lazy('ticket:index')
-
-    def form_valid(self, form):
-        seat_number = form.cleaned_data.get('seat_number')
-        if not seat_number:
-            form.add_error('seat_number', 'Seat must be selected')
-            return self.form_invalid(form)
-        return super().form_valid(form)
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data()
-        return context
-
-
-class TicketUpdate(UpdateView):
-    model = Ticket
-    form_class = CreateTicketForm
-    template_name = 'ticket/update.html'
-    success_url = reverse_lazy('ticket:update')
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        ticket = self.object
-
-        context['form'].fields['exam_type'].initial = ticket.exam_type
-        context['form'].fields['precinct'].initial = ticket.precinct
-        context['form'].fields['floor'].initial = ticket.floor
-        context['form'].fields['room'].initial = ticket.room
-        context['form'].fields['seat_number'].initial = ticket.seat_number
-
-        return context
-
-
+@login_required
 def get_ticket(request, ticket_id):
     ticket = Ticket.objects.get(id=ticket_id)
     context = {
@@ -127,12 +141,25 @@ def get_ticket(request, ticket_id):
     return render(request, 'ticket/ticket.html', context=context)
 
 
+@login_required
 def delete_ticket(request, ticket_id):
-    ticket = Ticket.objects.get(id=ticket_id)
-    ticket.delete()
-    return render(request, 'ticket/index.html')
+    if request.method == "POST":
+        ticket = get_object_or_404(Ticket, id=ticket_id)
+
+        # Увеличиваем количество свободных мест в комнате
+        room = ticket.room
+        room.available_seats += 1
+        room.save()
+
+        # Удаляем билет
+        ticket.delete()
+
+        return JsonResponse({"message": "Uğurla silindi"}, status=200)
+
+    return JsonResponse({"error": "Yanlış sorğu metodu"}, status=400)
 
 
+@login_required
 def get_report(request):
     exams = Exam.objects.all()
     context = {
@@ -141,12 +168,14 @@ def get_report(request):
     return render(request, 'ticket/report.html', context=context)
 
 
+@login_required
 def load_exam_type_report(request):
     exam = request.GET.get('exam')
     exam_typs = ExamType.objects.filter(exam=exam).values('id', 'name')
     return JsonResponse(list(exam_typs), safe=False)
 
 
+@login_required
 def export_to_excel(request):
     exam_id = request.GET.get('exam')
     exam_type_id = request.GET.get('exam_type')
@@ -192,3 +221,120 @@ def export_to_excel(request):
     workbook.save(response)
 
     return response
+
+
+# EXAM
+class ExamListView(LoginRequiredMixin, ListView):
+    model = Exam
+    template_name = 'ticket/exam/exam-list.html'
+    context_object_name = 'exams'
+
+    def get_queryset(self):
+        return Exam.objects.filter(user=self.request.user)
+
+    def get_context_data(self, *, object_list=None, **kwargs):
+        context = super().get_context_data()
+        return context
+
+
+class ExamCreateView(LoginRequiredMixin, CreateView):
+    model = Exam
+    form_class = CreateExamForm
+    template_name = 'ticket/exam/add-exam.html'
+    success_url = reverse_lazy('ticket:index')
+
+    def form_valid(self, form):
+        form.instance.user = self.request.user
+        return super().form_valid(form)
+
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data()
+        return context
+
+
+# PRECINCT
+class PrecinctListView(LoginRequiredMixin, ListView):
+    model = Precinct
+    template_name = 'ticket/exam/precinct-list.html'
+    context_object_name = 'precincts'
+
+    def get_queryset(self):
+        """Фильтруем Precinct по текущему пользователю и считаем этажи"""
+        return Precinct.objects.filter(user=self.request.user).annotate(floor_count=Count('floor'))
+
+    def get_context_data(self, *, object_list=None, **kwargs):
+        context = super().get_context_data()
+        return context
+
+
+class CreatePrecinctView(LoginRequiredMixin, CreateView):
+    model = Precinct
+    form_class = CreatePrecinctForm
+    template_name = 'ticket/exam/add-precinct.html'
+    success_url = reverse_lazy('ticket:precinct_list')
+
+    def get_form_kwargs(self):
+        """Передаем текущего пользователя в форму"""
+        kwargs = super().get_form_kwargs()
+        kwargs['user'] = self.request.user  # Передаем user
+        return kwargs
+
+    def form_valid(self, form):
+        """Привязываем Precinct к пользователю"""
+        form.instance.user = self.request.user
+        return super().form_valid(form)
+
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data()
+        return context
+
+
+def delete_precinct(request, precinct_id):
+    if request.method == 'POST':
+        precinct = Precinct.objects.get(pk=precinct_id)
+
+        precinct.delete()
+        return JsonResponse({"message": "Uğurla silindi"}, status=200)
+
+    return JsonResponse({"error": "Yanlış sorğu metodu"}, status=400)
+
+
+# FLOOR
+
+
+class FloorListView(ListView):
+    model = Floor
+    template_name = 'ticket/exam/floor-list.html'  # Шаблон для отображения этажей
+    context_object_name = 'floors'
+
+    def get_queryset(self):
+        """Фильтруем этажи (Floor) только по выбранному Precinct"""
+        precinct_id = self.kwargs.get('precinct_id')
+        return Floor.objects.filter(precinct_id=precinct_id)
+
+    def get_context_data(self, **kwargs):
+        """Добавляем объект Precinct в контекст, чтобы вывести его в шаблоне"""
+        context = super().get_context_data(**kwargs)
+        context['precinct'] = get_object_or_404(Precinct, id=self.kwargs.get('precinct_id'))
+        return context
+
+
+class FloorCreateView(LoginRequiredMixin, CreateView):
+    model = Floor
+    form_class = FloorCreateForm
+    template_name = 'ticket/exam/add-floor.html'
+
+    def get_form_kwargs(self):
+        """Передаем текущего пользователя в форму"""
+        kwargs = super().get_form_kwargs()
+        kwargs['user'] = self.request.user  # Передаем user
+        return kwargs
+
+    def form_valid(self, form):
+        """Привязываем Precinct к пользователю"""
+        form.instance.user = self.request.user
+        return super().form_valid(form)
+
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data()
+        return context
